+++ src/Common/xrpl-network.ts
@@ -1,5 +1,16 @@
 /**
  * XRP Ledger Networks.
+ *
+ * @deprecated Use XrplNetwork.
+ */
+export enum XRPLNetwork {
+  Dev = 'devnet',
+  Test = 'testnet',
+  Main = 'mainnet',
+}
+
+/**
+ * XRP Ledger Networks.
  */
 enum XrplNetwork {
   Dev = 'devnet',